fix: broadcast limits, WS invalidations, point overflow
 - Cap teams, bases, and locations to 500 in broadcast payloads
 - Skip location data for non-live games
 - Batch React Query invalidations via requestAnimationFrame in both WS hooks
 - Change leaderboard points from int to long to prevent overflow
 - Restrict CORS allowed headers to Authorization + Content-Type
 - Wrap RichTextEditor in ErrorBoundary on ChallengesPage

+++ backend/src/main/java/com/prayer/pointfinder/dto/response/LeaderboardEntry.java
@@ -13,6 +13,6 @@ public class LeaderboardEntry {
     private UUID teamId;
     private String teamName;
     private String color;
-    private int points;
+    private long points;
     private int completedChallenges;
 }
